# perf(bencode): sync.Map for struct field caching to reduce contention