Replaced sync.Mutex and sync.RWMutex with sync.Map in decode/encode struct caching. This optimizes for read-heavy concurrent access by avoiding lock overhead, reducing execution time without increasing memory usage. Benchmarks below:

**Before**
```
encode_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/bencode
cpu: Apple M1 Pro
BenchmarkGetEncodeFields-10      7381923           167.3 ns/op         0 B/op          0 allocs/op
PASS
ok      github.com/james-lawrence/torrent/bencode   1.636s

decode_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/bencode
cpu: Apple M1 Pro
BenchmarkGetStructFieldForKey-10         8815094           138.8 ns/op         0 B/op          0 allocs/op
PASS
ok      github.com/james-lawrence/torrent/bencode   1.588s
```

**After**
```
encode_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/bencode
cpu: Apple M1 Pro
BenchmarkGetEncodeFields-10     1000000000           1.065 ns/op           0 B/op          0 allocs/op
PASS
ok      github.com/james-lawrence/torrent/bencode   1.433s 

decode_test.go
goos: darwin
goarch: arm64
pkg: github.com/james-lawrence/torrent/bencode
cpu: Apple M1 Pro
BenchmarkGetStructFieldForKey-10        445837135            2.583 ns/op           0 B/op          0 allocs/op
PASS
ok      github.com/james-lawrence/torrent/bencode   1.673s
```